// src/pages/type2/videomenueslisting/models/videoedit.dto.ts
export interface VideoUpdate {
    parentclub_id: string
   activity_id: string
    ParentClubKey: string
    ClubKey: string
    MemberKey: string
    AppType: number
    ActionType: number
    DeviceType: number
    Id: string
    activityKey: string
    is_headline: boolean
    video_description: string
    video_tag: string
    video_title: string
    video_url: string
    thumbnail: string
    associated_activity: string
    is_show_applus: boolean
    is_show_member: boolean
    unsanitized_video_url: string
    }